+++ tests/test_run_case_list.py
@@ -2,7 +2,6 @@ import datetime
 import pathlib
 import uuid
 
-import cattr
 import requests_mock
 
 from varfish_cli.__main__ import main
@@ -11,13 +10,13 @@ from varfish_cli.api import models, CONVERTER
 
 def test_run_case_list():
     url = "https://varfish.example.com"
-    address = "%s/variants/api/case/c3752df7-fa32-4784-8a48-e8f0e5a28790/" % url
+    address = "%s/variants/api/case/list/c3752df7-fa32-4784-8a48-e8f0e5a28790/" % url
     cases = CONVERTER.unstructure(
         [
             models.Case(
                 sodar_uuid=str(uuid.uuid4()),
-                date_created=datetime.datetime.now().isoformat(),
-                date_modified=datetime.datetime.now().isoformat(),
+                date_created=datetime.datetime.now(),
+                date_modified=datetime.datetime.now(),
                 name="Case_Name",
                 index="index",
                 pedigree=[